🎊完善网关，增加登录拦截器，有bug

+++ geekedu-common/common-satoken/src/main/java/com/zch/common/satoken/context/UserContext.java
@@ -24,7 +24,7 @@ public class UserContext {
     }
 
     public static Long getLoginId() {
-        return (Long) getThreadLocalMap().get("loginId");
+        return getThreadLocalMap().get("loginId");
     }
 
     public static void remove() {
